Pusatkan konfigurasi & hapus kredensial hardcoded dari kode
- Tambah config.example.php (template) & midtrans_config.php (bootstrap)
- Pindahkan key Midtrans/DB ke config.php (di-gitignore)
- Hapus Server/Client Key Midtrans yang sebelumnya hardcoded
- Tambah admin/admin_guard.php & pengetatan keamanan halaman admin

+++ detail_chat.php
@@ -8,7 +8,7 @@ $is_user     = is_user_login();
 $is_designer = is_designer_login();
 
 // Ambil ID Pelaku
-$id_saya = current_id();
+$id_saya = (int) current_id();
 
 // 2. TANGKAP ID LAWAN BICARA
 if (isset($_GET['tujuan'])) {
